==> notebooks/4_0_API_Predictions.ipynb <==
! pip install requests

import requests

BASE_URL = "https://etf-stock.onrender.com/api"
PREDICTION_ENDPOINT = lambda vol_moving_avg, adj_close_rolling_med, model_type="Random Forest":f"predict?vol_moving_avg={vol_moving_avg}&adj_close_rolling_med={adj_close_rolling_med}&model_type={model_type}"

vol_moving_avg = 2.9
adj_close_rolling_med = 22224.0

def make_prediction(vol_moving_avg, adj_close_rolling_med):
    url = f"{BASE_URL}/{PREDICTION_ENDPOINT(vol_moving_avg, adj_close_rolling_med)}"
    response = requests.get(url=url)
    if response.status_code == 200:
        return int(response.content.decode(encoding='utf-8'))
    else:
        print(f"url={url} Status code: {response.status_code}")
        return None

prediction = make_prediction(vol_moving_avg, adj_close_rolling_med)
print(f"Prediction: {prediction}")
